requirements.txt
amqp==2.4.2
Babel==2.6.0
billiard==3.5.0.5
bootstrap4==0.1.0
celery==4.2.1
certifi==2019.3.9
chardet==3.0.4
crontab==0.22.5
dataclasses==0.6
Django==2.1.7
django-constance==2.4.0
django-debug-toolbar==1.11
django-filter==2.1.0
django-filters==0.2.1
django-jalali==3.1.0
django-js-asset==1.2.2
django-model-utils==3.1.2
django-mptt==0.9.1
django-multiselectfield==0.1.8
django-sendgrid-webhook==0.3.0
django-stubs==0.9.0
djangorestframework==3.9.2
djangorestframework-stubs==0.3.0
drfpasswordless==1.3.7
flower==0.9.3
idna==2.8
jdatetime==3.2.0
kombu==4.4.0
mypy==0.670
mypy-extensions==0.4.1
Pillow==5.4.1
psycopg2==2.7.7
PyJWT==1.7.1
PySocks==1.6.8
pytz==2018.9
redis==3.2.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
tornado==5.1.1
twilio==6.25.2
typed-ast==1.3.1
typing-extensions==3.7.2
urllib3==1.24.1
vine==1.3.0
